feat(classes): Move classes
