--- app/controllers/project-version/class/events.js
@@ -1,1 +1,0 @@
-export { default } from '../../events';
