--- app/controllers/project-version/class/methods.js
@@ -1,1 +1,0 @@
-export { default } from '../../methods';
