--- app/controllers/project-version/class/properties.js
@@ -1,1 +1,0 @@
-export { default } from '../../properties';
