+++ app/controllers/project-version/classes/class/events.js
@@ -1,0 +1,1 @@
+export { default } from '../../../events';
