+++ app/controllers/project-version/classes/class/methods.js
@@ -1,0 +1,1 @@
+export { default } from '../../../methods';
