+++ app/controllers/project-version/classes/class/properties.js
@@ -1,0 +1,1 @@
+export { default } from '../../../properties';
