--- app/controllers/project-version/module/events.js
@@ -1,1 +1,0 @@
-export { default } from '../../events';
