--- app/controllers/project-version/module/methods.js
@@ -1,1 +1,0 @@
-export { default } from '../../methods';
