--- app/controllers/project-version/module/properties.js
@@ -1,1 +1,0 @@
-export { default } from '../../properties';
