+++ app/controllers/project-version/modules/module/events.js
@@ -1,0 +1,1 @@
+export { default } from '../../../events';
