+++ app/controllers/project-version/modules/module/methods.js
@@ -1,0 +1,1 @@
+export { default } from '../../../methods';
