+++ app/controllers/project-version/modules/module/properties.js
@@ -1,0 +1,1 @@
+export { default } from '../../../properties';
