+++ app/controllers/project-version/namespaces/namespace/index.js
@@ -1,3 +1,3 @@
-import ClassIndexController from "../../class/index";
+import ClassIndexController from "../../classes/class/index";
 
 export default ClassIndexController.extend();
