+++ app/routes/project-version/classes/class/properties.js
