+++ app/templates/project-version/modules/module.hbs
@@ -33,7 +33,7 @@
       <ul class="spec-property-list">
         {{#each classesAndNamespaces as |klass|}}
           <li>
-            {{#link-to 'project-version.class' klass}}
+            {{#link-to 'project-version.classes.class' klass}}
               {{klass}}
             {{/link-to}}
           </li>
